Final update with Try
Okay, I know I got a little out of turn here, but I got so excited with my or statements, I wrote the logic to account for if each of the values were negative.  So, I'm here adding the testing for the 3rd score after writing the logic (not TDD, I know).

Also, I added a test to try all 3 negative, and it still passes.