Selection Using Dictionary
Commit 2: this is all we need.

I reeeeeallly didn't want to commit after passing each test individually.  They all passed the first time.  This was super straightforward.  I was a little confused, however, on whether or not I should be passing in a value, or not.  I ended up not.